requirements.txt
python-dotenv
loguru
vk_api